finally understand defaults, url, and working with fetch
However, when editing a subtest the subtest only knows what the ID of
the assessment it belongs to. I think maybe a simple fetch to get the
name and get rid of superFetch, which uses _id, instead of name. seems
costly, wish we could throw around objects without risking memory leaks